requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
idna==3.4
minio==7.2.0
pycparser==2.21
pycryptodome==3.19.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
urllib3==2.0.4
